requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
discord.py @ git+https://github.com/Rapptz/discord.py@e0341c9ef25c0093bff991ff09c87dba0414e58c
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
requests==2.28.1
urllib3==1.26.10
yarl==1.7.2
